GHAppy

Running GHAppy.

GHAppy is meant to be easy to work with, and these two functions together with the exposed API should be all you need to come up with a nice report. To see an example of it in action, please refer to 'src/Main.hs'.

runGHAppy :: Settings -> Eff (GHAppyAct ': GHStack) a -> IO a

runCompose :: forall effs a. EffGH effs => Eff (Composer ': effs) a -> Eff effs [Leaf]

The parameters of GHAppy.

data Settings

Used by GHAppy for necessary information.

Constructors

Settings

Fields

apiKey :: String

outputDirectory :: String

outputFile :: String

repository :: String

userAgent :: String

pandocTemplateUrl :: String

preambleLocation :: FilePath

logEnvironment :: LogEnv

GHAppy Effect, and Effect Stacks.

type GHStack = '[Logger, Reader Settings, State Issues, IO]

The usual Eff Stack for GHAppy.

data Logger a where

Logger is an Eff that allows us to log, in the default implementation Katip is used for structured logging.

Constructors

LogS
Standard Info Log.

Fields

:: [String]

Log Environment

-> String

Log Message

-> Logger ()

Log Effect

data GHAppyAct a where

GHAppy API.

Constructors

SetUpDir :: GHAppyAct ()
Creates a directory as specified by Settings.

PullIssues :: GHAppyAct Issues
Pulls all the Issues from the repository specified in Settings, and keeps them in memory.

SaveAvailableIssues :: GHAppyAct [FilePath]
Dumps all the issues from memory into a specific folder.

GeneratePDF :: [Leaf] -> GHAppyAct ()
Uses Pandoc to generate the final report.

data Composer a where

API for composing documents from issues.

Constructors

AddFile :: Integer -> IssueN -> Composer ()
Adds the Issue with a specific number to the composer.

AddNewPage :: Composer ()
Adds an empty page.

AddAllPagesThat :: Integer -> Predicate Issue -> Composer ()
Adds all the issues that satisfy a predicate, bumping their headers by a specific amount.

AddHeader :: Integer -> String -> Composer ()
Adds a header at specific level.

AddDisclaimer :: Composer ()
Adds the standard disclaimer.

AddVulnTypes :: Composer ()
Adds the standard

data Issue

Issue is the standard document GHAppy works with. It is a simpler version of Entry and the reasoning behind decoupling the two is both a matter of history and convenience. In future versions this could be replaced with Entry directly.

type EffGH a = (Members GHStack a, LastMember IO a)

A constraint that includes the GHSTack and has IO as the last Eff.

GHappy specific types.

data Issues

A wrapper around a map of IssueN and Issue. It is used for the in-memory representation of Issues.

type IssueN = Integer

The number of an Issue, as found on GitHub.

data Leaf

A Leaf is to compose the final output.

Constructors

Leaf

Fields

issueN :: Maybe IssueN

Leafs can contain an IssueN - in which case the content of the Issue is included in the final output. Alternatively they can simply be in-place text, in which case only the Leaf's preamble is included.

preamble :: [String]

The Preamble is prepended to an included leaf. We use this to either prepend some custom text to an issue, or to create a non issue entry.

level :: Integer

The level of a leaf represents by how many levels we will bump the headers it contains. '0' means none.

GHAppy's logistical API.

setUpDir :: forall effs. Member GHAppyAct effs => Eff effs ()

pullIssues :: forall effs. Member GHAppyAct effs => Eff effs Issues

saveAvailableIssues :: forall effs. Member GHAppyAct effs => Eff effs [FilePath]

generatePDF :: forall effs. Member GHAppyAct effs => [Leaf] -> Eff effs ()

Composer specific API.

addAllPagesThat :: forall effs. Member Composer effs => Integer -> Predicate Issue -> Eff effs ()

addNewPage :: forall effs. Member Composer effs => Eff effs ()

addHeader :: forall effs. Member Composer effs => Integer -> String -> Eff effs ()

addFile :: forall effs. Member Composer effs => Integer -> IssueN -> Eff effs ()

addDisclaimer :: forall effs. Member Composer effs => Eff effs ()

addVulnTypes :: forall effs. Member Composer effs => Eff effs ()

GHAppy predicates.

Predicates are an easy way to create filters for the issues you would want to include in the report. It is the recommended way to batch the retrieval and inclusion of Issues, rather than the inclusion by number. The concatenation Predicates provides an intuitive way of composing them.

hasOnlyLabel :: String -> Predicate Issue

Predicate for issues that have only a specific label.

hasLabel :: String -> Predicate Issue

Predicate for issues that have a specific label.

isOpen :: Predicate Issue

Predicate for open issue.

Logger API.

logS :: forall effs. Member Logger effs => [String] -> String -> Eff effs ()