GHAppy.Types

Documentation

newtype Label

Labels (in addition to issue numbers) are the main way of filtering and selecting issues to be included in the final report.

Constructors

Label

Fields

name :: Text

data Entry

An Entry is a GH Issue. We do not require all the information for the purpose of GHAppy, rather only a subset.

Constructors

Entry

Fields

number :: Integer

title :: Text

labels :: [Label]

state :: String

body :: Maybe Text